requirements.txt
# FastAPI 관련
fastapi>=0.115.0
uvicorn>=0.22.0
pydantic>=2.10.6  # 최신 버전 반영
jinja2>=3.1.3  # 템플릿 엔진
aiofiles>=23.2.1  # 비동기 파일 처리

# Google Maps 관련
googlemaps>=4.10.0  # 최신 버전 반영

# 데이터베이스 관련 (프로젝트에 따라 선택)
sqlalchemy>=2.0.37
pymysql>=1.1.1  # MySQL을 사용하는 경우
psycopg2-binary>=2.9.10  # PostgreSQL을 사용하는 경우

# 환경변수 관리
python-dotenv>=1.0.1  # 최신 버전 반영

# 기타 유틸리티
requests>=2.32.3
python-multipart>=0.0.20  # 최신 버전 반영

# YouTube 관련
youtube-transcript-api>=0.6.1
google-api-python-client>=2.108.0
pytube>=15.0.0

# LangChain 관련
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.18
tiktoken>=0.5.2

# 텍스트 처리
beautifulsoup4>=4.12.2
langdetect>=1.0.9